refactor: sync order of public handlers, tests and API docs
